requirements.txt
alembic==1.17.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Cors==4.0.0
flask-marshmallow==1.3.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
packaging==25.0
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
